Move package to dev.drewhamilton.androidtime
