+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="dev.drewhamilton.androiddatetimeformatters.demo">
+    package="dev.drewhamilton.androidtime.format.demo">
 
   <application
       android:name=".DemoApp"
