+++ javatime/src/androidTest/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.drewhamilton.androiddatetimeformatters.javatime.test">
+    package="dev.drewhamilton.androidtime.format.test">
 
   <uses-permission android:name="android.permission.WRITE_SETTINGS" />
