+++ test/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.drewhamilton.androiddatetimeformatters.test" />
+    package="dev.drewhamilton.androidtime.format.test" />
